+++ cameo-api/CHANGELOG.md
@@ -14,6 +14,7 @@
 * Added Instance.waitFor(KeyValue) to wait for a key value event.
 * Added Com classes for port and storage requests.
 * Argument info passed in JSON format.
+* Added optional stopping time in This.handleStop() method.
 
 0.1.9
 -----